Switch "delete target on failed update?" test to protect targets marked with 'actions updated', not 'actions together'.
The idea is that targets with state (those updated, not replaced)
shouldn't be deleted if their update fails.  But the test was on
'actions together' (which in the stock Jambase is always used with
'actions updated'), when it should be on 'actions updated'.

Bug fix documented in RELNOTES.

Jobs: job000013
[git-p4: depot-paths = "//guest/perforce_software/jam/src/": change = 1494]