A few more amd64 build changes.  Remove conflicting /MT /MTd flags. Remove common7 from jamsh.bat PATH setting.
As a note, built a 64bit p4d.exe without smartheap.  This port
is going to be some work.  For starters, _get_osfhandle() now
has checks which fault the server.  It's not clear on what it
will take to resolve this.  I suspect MS has a bunch more issues
like this which must be addressed for the port.


p4transfer.py: Transferred from production

[git-p4: depot-paths = "//guest/perforce_software/jam/src/": change = 9860]