On a Windows debug build the /MT flag was used in addition to the /MTd flag.  Basically harmless except the compiler issued a warning for every file compiled.
Removed the use of the /MT flag in the jambase arena.  This
allows the Jamrules to properly set this flag as required.

Also when doing a Smart Heap debug build, duplicate symbols
in the Smart Heap library conflicted with the standard
Windows CRT library.  Adding a /FORCE to the link line solves
this problem.  This is a very narrow case and should not effect
builds without Smart Heap.

Jam must be rebuilt to remove the /MT and /MTd warning.


p4transfer.py: Transferred from production

[git-p4: depot-paths = "//guest/perforce_software/jam/src/": change = 9947]